+++ fast64_internal/oot/scene/exporter/to_c/room_shape.py
@@ -1,7 +1,6 @@
-from .....utility import CData, PluginError
+from .....utility import CData, PluginError, indent
 from .....f3d.f3d_gbi import ScrollMethod, TextureExportSettings
 from ....oot_f3d_writer import OOTGfxFormatter
-from ....oot_utility import indent
 from ....oot_constants import ootRoomShapeStructs, ootRoomShapeEntryStructs, ootEnumRoomShapeType
 from ....oot_level_classes import OOTRoom, OOTRoomMeshGroup, OOTRoomMesh
